% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast_primers}
\alias{blast_primers}
\title{Blast primers}
\usage{
blast_primers(primer, ref, path_out = NULL, format = 7)
}
\arguments{
\item{primer}{Path to fasta file with primer sequences.}

\item{ref}{Path to fasta file with reference.}

\item{path_out}{Path to directory to export files.}

\item{format}{Blast format. Defaults to \code{7}.}
}
\description{
Blast primers against reference sequence
}
