% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blast_sequence}
\alias{blast_sequence}
\title{Blast sequence}
\usage{
blast_sequence(
  seq,
  ref,
  what = c("gene", "primer"),
  format = 7,
  path_out = NULL,
  seq_name = NULL,
  ref_name = NULL
)
}
\description{
Blast sequence
}
