% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gwas.R
\name{calculate_maf}
\alias{calculate_maf}
\title{Calculate MAF}
\usage{
calculate_maf(bed, rs)
}
\arguments{
\item{bed}{Path to bed file.}

\item{rs}{String specifying the target SNP (from column RS in the output from gemma).}
}
\description{
Calculate minor allele frequency
}
\details{
Uses \code{bigsnpr} to pull the SNP specified by \code{rs} and calculates the
minor allele frequency by dividing the number of individuals with the minor
minor allele by the total number of individuals.
}
