% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{cluster_blast}
\alias{cluster_blast}
\title{Cluster blast}
\usage{
cluster_blast(df, max_gap = 5000, min_elements = 1)
}
\arguments{
\item{df}{Output from \code{\link{read_blast}}.}

\item{max_gap}{Maximum gap between hits for clustering. Defaults to
\code{5000}.}

\item{min_elements}{The mimimum number of elements in a cluster. Clusters
with fewer elements will be filtered out. Defaults to \code{1}.}
}
\description{
Cluster blast results
}
\details{
This is useful for grouping together exons that belong to the same gene.
}
