% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_genes.R
\name{extract_genes}
\alias{extract_genes}
\title{Extract genes}
\usage{
extract_genes(
  genes,
  fasta,
  gff3 = NULL,
  path_out = NULL,
  filenames = c("id", "name"),
  what = c("gene", "cds", "exons"),
  translate = FALSE
)
}
\arguments{
\item{genes}{GFF annotation file as \code{\link[data.table:data.table]{data.table}}.}

\item{fasta}{Path to fasta file.}

\item{gff3}{Path to GFF3 file or \code{data.table} generated by \code{\link{read_gff}}./}

\item{path_out}{Path to export fasta files for the specified genes.}

\item{filenames}{Whether to use \code{id} or \code{name} as filenames for the
exported files.}

\item{what}{What to export: either \code{gene}, \code{cds}, or \code{exons}.}

\item{translate}{Whether to translate the DNA sequence (if \code{what == "exons"}).}
}
\description{
Extract gene sequences based on GFF3 annotation file.
}
