% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snps.R
\name{extract_snps}
\alias{extract_snps}
\title{Extract SNPs from BED file by ID}
\usage{
extract_snps(bed, snp_ids)
}
\arguments{
\item{bed}{Path to BED file.}

\item{snp_ids}{Vector of SNP IDs.}
}
\description{
Extract SNPs from BED file by ID
}
\author{
Ethan Bass
}
