% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_parsing.R, R/parse_gff.R
\name{get_children}
\alias{get_children}
\title{Get exons from parent}
\usage{
get_children(x, gff, what = c("CDS", "exon"))

get_children(x, gff, what = c("CDS", "exon"))
}
\description{
Get exons from parent

Get exons from parent
}
