% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes_by_idx}
\alias{get_genes_by_idx}
\title{Get genes by index}
\usage{
get_genes_by_idx(df, genes, idx = 1, half_width = 50000)
}
\arguments{
\item{genes}{GFF annotation file as \code{data.table}.}

\item{half_width}{Half width. Defaults to \code{50000}.}
}
\description{
Get genes by index
}
