% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes_by_pos}
\alias{get_genes_by_pos}
\title{Get genes by position}
\usage{
get_genes_by_pos(genes, chr, loc, half_width = 50000)
}
\arguments{
\item{genes}{GFF annotation file as data.table}

\item{chr}{String denoting chromosome}

\item{loc}{String denoting position on chromosome}

\item{half_width}{Half width. Defaults to \code{50000}.}
}
\description{
Get genes by position
}
