% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_parsing.R, R/parse_gff.R
\name{get_mRNA}
\alias{get_mRNA}
\title{Get mRNA
Get mRNA from gene parent}
\usage{
get_mRNA(x, gff)

get_mRNA(x, gff)
}
\description{
Get mRNA
Get mRNA from gene parent

Get mRNA
Get mRNA from gene parent
}
