% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_peaks}
\alias{get_peaks}
\title{Get peaks}
\usage{
get_peaks(df, genes, n = NULL, t = NULL, half_width = 1000)
}
\arguments{
\item{genes}{GFF annotation file as \code{data.table}.}
}
\description{
Get peaks
}
