% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snps.R
\name{get_snps}
\alias{get_snps}
\title{Extract SNPs from BED file by ID}
\usage{
get_snps(bed, snp_ids, fam_cols = "FID", rename = TRUE)
}
\arguments{
\item{bed}{Path to BED file.}

\item{snp_ids}{Vector of SNP IDs.}

\item{fam_cols}{Which "family" columns to include in the output. Options are
\code{FID} and/or \code{"IID"}.}

\item{rename}{Logical. Whether to simplify the names of the SNPs. Defaults to
\code{TRUE}.}
}
\description{
Extracts raw SNP data from a \code{bed} file for SNPs specified by
\code{snp_ids}.
}
\author{
Ethan Bass
}
