% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gwas.R
\name{gff_2gggenes}
\alias{gff_2gggenes}
\title{Convert GFF (general feature format) to 'gggenes' format}
\usage{
gff_2gggenes(genes, chr, pos)
}
\arguments{
\item{genes}{GFF annotation file as data.table}

\item{chr}{Chromosome}

\item{pos}{Position on chromosome}
}
\description{
Convert GFF (general feature format) to 'gggenes' format
}
