% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{plot_blast}
\alias{plot_blast}
\title{Plot blast results}
\usage{
plot_blast(path, max_gap = 1e+05)
}
\arguments{
\item{path}{Path to blast results.}

\item{max_gap}{Max gap between blast hits for clustering. Defaults to
\code{100000}.}
}
\description{
Plot blast results
}
