% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{plot_conservation_primers}
\alias{plot_conservation_primers}
\title{Plot primer conservation}
\usage{
plot_conservation_primers(primerF, primerR, seqs, conservation)
}
\description{
Plot primer conservation
}
