% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_parsing.R, R/parse_gff.R
\name{plot_genes}
\alias{plot_genes}
\title{Plot genes
Plot genes from GFF annotations}
\usage{
plot_genes(
  gff,
  what = c("exons", "genes"),
  legend_col = 2,
  legend_size = 0.5,
  feature_width = 1.2
)

plot_genes(
  gff,
  what = c("exons", "genes"),
  legend_col = 2,
  legend_size = 0.5,
  feature_width = 1.2
)
}
\description{
Plot genes
Plot genes from GFF annotations

Plot genes
Plot genes from GFF annotations
}
