% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fasta.R, R/plot_gwas.R
\name{plot_gwas}
\alias{plot_gwas}
\title{Plot fasta}
\usage{
plot_gwas(
  x,
  gff,
  bed,
  G,
  threshold = 7.5,
  thr.r2 = 0.2,
  size = 100/thr.r2,
  ncores = 2,
  max_hits = 21,
  nrow = 4,
  n_genes = 10,
  title = TRUE,
  ...
)

plot_gwas(
  x,
  gff,
  bed,
  G,
  threshold = 7.5,
  thr.r2 = 0.2,
  size = 100/thr.r2,
  ncores = 2,
  max_hits = 21,
  nrow = 4,
  n_genes = 10,
  title = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{data.table} of GWAS hits.}

\item{gff}{GFF annotation file as \code{\link[data.table:data.table]{data.table}}.}

\item{bed}{Path to BED file.}

\item{G}{SNPs in \code{bigsnpr} format or path to BED file.}

\item{threshold}{Significance threshold for SNPs to include (-log10(p)).}

\item{thr.r2}{Argument to Threshold over the squared correlation between two SNPs. Default is 0.2.}

\item{size}{For one SNP, window size around this SNP to compute correlations.}

\item{ncores}{Number of cores to use for clumping. Defaults to \code{2}.}

\item{max_hits}{Number of SNPs to plot.}

\item{nrow}{Number of genes to plot per page.}

\item{n_genes}{Maximum number of genes to plot per SNP.}

\item{title}{Whether to include titles.}

\item{...}{Additional arguments to \code{plot_gwas_single}.}
}
\value{
Invisibly returns list of individual plots.
}
\description{
Plot gene arrow diagram from fasta headers

Plots gene arrow diagrams for the top SNPs showing genes associated with each
SNP after
}
\details{
Plots gene arrow diagrams for the locations specified in fasta header
}
\author{
Ethan Bass
}
