% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot}
\usage{
plot_manhattan(
  x,
  chr = "chr",
  bp = "ps",
  p = NULL,
  n = NULL,
  title = NULL,
  ...
)
}
\arguments{
\item{x}{A \code{data.frame} containing the GWAS results.}

\item{chr}{A string denoting the name of the chromosome column in \code{x}.
Defaults to \code{"chr"} corresponding to the output from gemma. Argument to
\code{fastman_gg}.}

\item{bp}{#' @param chr A string denoting the column name for the chromosomal
position in \code{x}. Defaults to \code{"ps"} corresponding to the output
from gemma. Argument to \code{fastman_gg}.}

\item{p}{#' @param chr A string denoting the column name for the p-value in
\code{x}. Defaults to \code{"p_lrt"} corresponding to the output from gemma.
Argument to \code{fastman_gg}.}

\item{n}{The number of SNPs for calculation of Bonferroni threshold. Defaults to
\code{nrow(x)}.}

\item{title}{A string denoting the title to be printed at the top of the plot.}

\item{...}{Additional arguments to \code{\link[fastman:fastman]{fastman::fastman_gg}}.}
}
\description{
Wrapper around the \code{fastman_gg} function in the
\code{\link[fastman:fastman]{fastman}} package.
}
\author{
Ethan Bass
}
