% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{read_blast}
\alias{read_blast}
\title{Read blast results}
\usage{
read_blast(path)
}
\arguments{
\item{path}{Path to blast (format 7) output file.}
}
\description{
Read blast results from tabular (format 7) files.
}
