% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_susie.R
\name{run_susie}
\alias{run_susie}
\title{Function to run Sum of Single Effects (SuSiE) Regression}
\usage{
run_susie(
  Y,
  bed,
  genes,
  chr,
  locus,
  window = 1e+06,
  coverage = 0.95,
  plot_it = TRUE,
  verbose = getOption("verbose"),
  ...
)
}
\arguments{
\item{Y}{Vector of observed responses.}

\item{bed}{Path to BED annotation file.}

\item{genes}{Path to GFF3 annotations (or a \code{data.table} of annotations).}

\item{chr}{The chromosome that the SNP is on.}

\item{locus}{The position of the SNP on chromosome \code{chr}.}

\item{window}{Half-width window around \code{locus} for inclusion of SNPs.}

\item{coverage}{Coverage, argument to \code{\link[susieR:susie]{susie}}.}

\item{plot_it}{Logical. Whether to plot the results or not. Defaults to
\code{TRUE}.}

\item{verbose}{Logical. Whether to print messages to console.}

\item{...}{Additional arguments to \code{\link[susieR]{susieR}}.}
}
\description{
Wrapper for Sum of Single Effects (SuSiE) regression in the
\code{\link[susieR:susie]{susieR}} package.
}
\author{
Ethan Bass
}
