% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gwas.R
\name{plot_gwas}
\alias{plot_gwas}
\title{Plot gene arrow diagram from GWAS results}
\usage{
plot_gwas(
  x,
  gff,
  bed,
  G,
  threshold = 7.5,
  max_hits = 21,
  nrow = 7,
  n_genes = 10,
  title = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{data.table} of GWAS hits.}

\item{gff}{GFF annotation file as \code{\link[data.table:data.table]{data.table}}.}

\item{bed}{Path to BED file.}

\item{G}{SNPs in \code{bigsnpr} format or path to BED file.}

\item{threshold}{Significance threshold for SNPs to include (-log10(p)).}

\item{max_hits}{Number of SNPs to plot.}

\item{nrow}{Number of genes to plot per page.}

\item{n_genes}{Maximum number of genes to plot per SNP.}

\item{title}{Whether to include titles.}

\item{...}{Additional arguments to \code{plot_gwas_single}.}
}
\description{
Plots gene arrow diagrams for the top SNPs showing genes associated with each
SNP.
}
\author{
Ethan Bass
}
