% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gwas.R
\name{plot_gwas}
\alias{plot_gwas}
\title{Plot gene arrow diagram from GWAS results}
\usage{
plot_gwas(
  gff,
  chr,
  pos,
  n = 10,
  legend_col = 2,
  legend_size = 0.5,
  feature_width = 1.2
)
}
\arguments{
\item{gff}{GFF annotation file as \code{\link[data.table:data.table]{data.table}}.}

\item{chr}{Chromosome or contig to plot.}

\item{pos}{Position in chromosome or contig to plot.}

\item{n}{Number of genes to include. This will be the top \code{n} genes
closest to the provided SNP.}

\item{legend_col}{Number of columns to include in legend.}

\item{legend_size}{Size of legend (in cm). Defaults to \code{0.5}.}

\item{feature_width}{Line width for feature annotations. Defaults to \code{1.2}.}
}
\description{
Plot gene arrow diagram from GWAS results
}
\author{
Ethan Bass
}
