% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gwas.R
\name{plot_gwas}
\alias{plot_gwas}
\title{Plot gene arrows}
\usage{
plot_gwas(
  genes,
  chr,
  pos,
  n = 10,
  legend_col = 2,
  legend_size = 0.5,
  feature_size = 1.2
)
}
\arguments{
\item{genes}{GFF annotation file as data.table.}

\item{chr}{String denoting chromosome.}

\item{pos}{String denoting position on chromosome.}

\item{legend_size}{Legend size in cm. Defaults to \code{0.5}.}

\item{feature_size}{Defaults to \code{1.2}.}
}
\description{
Plot gene arrows
}
