% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gff.R
\name{read_gff}
\alias{read_gff}
\title{Read GFF3}
\usage{
read_gff(path, what = c("gene", "mRNA", "CDS", "exon", "all"))
}
\arguments{
\item{path}{Path to GFF3 file.}

\item{what}{What features to include (\code{gene} or \code{mRNA}).}
}
\description{
Read GFF3
}
\author{
Ethan Bass
}
